+++ lib/tasks.dart
@@ -1,0 +1,10 @@
+library hop_tasks;
+
+import 'dart:io' as io;
+import 'package:bot/bot.dart';
+import 'package:hop/hop.dart';
+import 'package:unittest/unittest.dart' as unittest;
+
+part 'src/tasks/unit_test.dart';
+part 'src/tasks/process.dart';
+part 'src/tasks/dart2js.dart';