+++ tool/hop_runner.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 import 'package:bot/bot.dart';
 import 'package:hop/hop.dart';
+import 'package:hop/tasks.dart';
 import '../test/console_test_harness.dart' as test_console;
 
 void main() {